+++ test/applyMiddleware.spec.js
@@ -1,8 +1,8 @@
 import expect from 'expect'
-import { createStore, applyMiddleware } from '../../src/index'
-import * as reducers from '../helpers/reducers'
-import { addTodo, addTodoAsync, addTodoIfEmpty } from '../helpers/actionCreators'
-import { thunk } from '../helpers/middleware'
+import { createStore, applyMiddleware } from '../src/index'
+import * as reducers from './helpers/reducers'
+import { addTodo, addTodoAsync, addTodoIfEmpty } from './helpers/actionCreators'
+import { thunk } from './helpers/middleware'
 
 describe('applyMiddleware', () => {
   it('wraps dispatch method with middleware once', () => {
@@ -69,7 +69,7 @@ describe('applyMiddleware', () => {
       {
         id: 1,
         text: 'Hello'
-      }, 
+      },
       {
         id: 2,
         text: 'World'
@@ -81,11 +81,11 @@ describe('applyMiddleware', () => {
         {
           id: 1,
           text: 'Hello'
-        }, 
+        },
         {
           id: 2,
           text: 'World'
-        }, 
+        },
         {
           id: 3,
           text: 'Maybe'
